+++ portal-impl/src/com/liferay/portal/lar/ThemeExporter.java
@@ -24,7 +24,7 @@ import com.liferay.portal.model.LayoutSet;
 import com.liferay.portal.model.LayoutSetBranch;
 import com.liferay.portal.model.Theme;
 import com.liferay.portal.model.adapter.ModelAdapterUtil;
-import com.liferay.portlet.layoutsadmin.lar.StagedTheme;
+import com.liferay.portal.model.adapter.StagedTheme;
 
 /**
  * @author Mate Thurzo
