+++ portal-impl/src/com/liferay/portal/model/adapter/builder/StagedGroupModelAdapterBuilder.java
@@ -15,8 +15,8 @@
 package com.liferay.portal.model.adapter.builder;
 
 import com.liferay.portal.model.Group;
-import com.liferay.portlet.sitesadmin.lar.StagedGroup;
-import com.liferay.portlet.sitesadmin.lar.StagedGroupImpl;
+import com.liferay.portal.model.adapter.StagedGroup;
+import com.liferay.portal.model.adapter.impl.StagedGroupImpl;
 
 /**
  * @author Mate Thurzo
