+++ portal-impl/src/com/liferay/portal/model/adapter/impl/StagedThemeImpl.java
@@ -12,11 +12,12 @@
  * details.
  */
 
-package com.liferay.portlet.layoutsadmin.lar;
+package com.liferay.portal.model.adapter.impl;
 
 import com.liferay.portal.kernel.lar.StagedModelType;
 import com.liferay.portal.kernel.util.StringPool;
 import com.liferay.portal.model.Theme;
+import com.liferay.portal.model.adapter.StagedTheme;
 import com.liferay.portal.model.impl.ThemeImpl;
 import com.liferay.portal.security.auth.CompanyThreadLocal;
 import com.liferay.portlet.expando.model.ExpandoBridge;
