+++ portal-impl/src/com/liferay/portlet/layoutsadmin/lar/StagedThemeStagedModelDataHandler.java
@@ -20,6 +20,7 @@ import com.liferay.portal.kernel.lar.PortletDataHandlerKeys;
 import com.liferay.portal.kernel.util.MapUtil;
 import com.liferay.portal.kernel.xml.Element;
 import com.liferay.portal.model.Theme;
+import com.liferay.portal.model.adapter.StagedTheme;
 import com.liferay.portal.service.ThemeLocalServiceUtil;
 
 import java.util.List;
