--- portal-service/src/com/liferay/portlet/layoutsadmin/lar/packageinfo
@@ -1,1 +1,0 @@
-version 1.0.0
